fix(web): merge team setup into team cards to remove duplication (ARC-427)
Each team was rendered twice: once as an editable row in TeamSetupPanel
(name input + color swatches + size stepper) and once as a static
header above its slots in TeamSlotsBoard. Now the editable controls
live inline at the top of each team's card (host-only); non-hosts see
the static colored team name. TeamSetupPanel reduces to the Add Team /
total slots / validation footer.
